+++ Class.h
@@ -44,20 +44,20 @@ class InfosRunner : public Runner {
 private:
     Runner michel;
     float vinst;
-    int traveled;
+    float traveled;
     int hydratation;
 public:
     InfosRunner();
     InfosRunner(Runner& newmichel,
         float& newvinst,
-        int& newtraveled,
+        float& newtraveled,
         int& newhydratation);
     Runner GetMichel();
     float GetVinst();
-    int GetTraveled();
+    float GetTraveled();
     int GetHydratation();
     void SetVinst(float vinst);
-    void SetTraveled(int traveled);
+    void SetTraveled(float traveled);
     void SetHydratation(int hydratation);
 
 };
